requirements.txt
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
